pyproject.toml
[project]
name = "dora-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "elevenlabs>=2.10.0",
    "gradio>=4.44.1",
    "groq>=0.31.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.5",
    "opencv-python>=4.12.0.88",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.14.3",
]
